move hooks to the custom hook
the PlayGround component was too big, so I moved hooks out to the
"useGameState" and declare them into the game, so now I have much less functions in the my main play ground
