+++ src/components/index.js
@@ -4,3 +4,4 @@ export { RandomSumIn } from "./RandomSumIn"
 export { DisplayStars } from "./DisplayStars"
 export { PlayAgain } from "./PlayAgain"
 export { GameTimer } from "./GameTimer"
+export { useGameState } from "./useGameState"
